requirements.txt
# --- Core ML & Data Science ---
torch
# Download appropritate gpu version of pytorch based on your pc specifications
# example isntall command: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu129   

sentence-transformers
scikit-learn
pydantic
numpy
pandas

# --- LightRAG & Dependencies ---
lightrag-hku[ollama,neo4j]
python-dotenv
faiss-cpu